// src/types/TimerTypes.ts
export type TimerHistoryType = {
  completed: boolean;
  completedRoutineCnt: number;
  createDate: string;
  id: number;
  routineCnt: number;
  setBreakTime: number;
  setFocusTime: number;
  taskName: string;
  totalTime: number;
};

export type TimerRecordType = {
  completed: boolean;
  completedRoutineCnt: number;
  createDate: string;
  id: number;
  routineCnt: number;
  setBreakTime: number;
  setFocusTime: number;
  totalTime: number;
  taskName: string;
};